pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "pydra-tasks-mriqc"
description = "Pydra tasks package for mriqc"
readme = "README.rst"
requires-python = ">=3.10"
dependencies = [
    "dipy",
    "fileformats >=0.8.3",
    "fileformats-datascience >=0.1",
    "fileformats-medimage >=0.4.1",
    "markupsafe ~= 2.0.1",
    "matplotlib",
    "nibabel",
    "nilearn",
    "migas >= 0.4.0",
    "pandas ~=1.0",
    "pydra >=0.23",
    "pydra-ants",
    "pydra-afni",
    "pydra-fsl",
    "pydra-mrtrix3 >=3.0.4a5",
    "pydra-niworkflows",
    "pydra-nireports",
    "PyYAML",
    "seaborn",
    "scikit-learn",
    "scipy",
    "statsmodels",
    "templateflow",
    "nilearn",
    "nitransforms",
    "torch",
    "toml",
    "tomli >= 1.1.0; python_version < '3.11'",
]
license = { file = "LICENSE" }
authors = [{ name = "Nipype developers", email = "neuroimaging@python.org" }]
maintainers = [
    { name = "Nipype developers", email = "neuroimaging@python.org" },
]
keywords = ["pydra"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["black", "pre-commit"]
doc = [
    "packaging",
    "sphinx >=2.1.2",
    "sphinx_rtd_theme",
    "sphinxcontrib-apidoc ~=0.3.0",
    "sphinxcontrib-napoleon",
    "sphinxcontrib-versioning",
]
test = [
    "nipype2pydra",
    "pytest >= 4.4.0",
    "pytest-cov",
    "pytest-env",
    "pytest-xdist",
    "pytest-rerunfailures",
    "codecov",
    "fileformats-extras",
    "fileformats-datascience-extras",
    "fileformats-medimage-extras",
    "fileformats-medimage-afni-extras",
    "fileformats-medimage-mrtrix3-extras",
    "fileformats-medimage-fsl-extras",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "pydra/tasks/mriqc/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["pydra"]
include-only = ["pydra/tasks/mriqc"]

[tool.black]
target-version = ["py38"]
exclude = "_version.py"

[tool.codespell]
ignore-words = ".codespell-ignorewords"

[tool.flake8]
doctests = true
per-file-ignores = ["__init__.py:F401,F403"]
max-line-length = 88
select = "C,E,F,W,B,B950"
extend-ignore = ['E203', 'E501', 'E129', 'W503']

[project.scripts]
# abide2bids = "mriqc.bin.abide2bids:main"
# dfcheck = "mriqc.bin.dfcheck:main"
# fs2gif = "mriqc.bin.fs2gif:main"
# mriqc = "mriqc.cli.run:main"
# mriqc_labeler = "mriqc.bin.labeler:main"
# mriqcwebapi_test = "mriqc.bin.mriqcwebapi_test:main"
# nib-hash = "mriqc.bin.nib_hash:main"
# participants = "mriqc.bin.subject_wrangler:main"
synthstrip = "pydra.tasks.mriqc.synthstrip.cli:main"
